CSL-404: resolve circular dependency issue.
